video should not be modified after build
